Update README and move a file
