+++ src/MOCK_DB_DATA/library-items/color/color-system.ts
@@ -1,4 +1,4 @@
-import { LibraryItemIds } from 'src/app/views/shared/library-item-ids';
+import { LibraryItemIds } from 'src/app/library-item-ids';
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
 import { BootstrapComponentsBaseUrl, NgBootstrapComponentsBaseUrl } from '../../bootstrap-base-urls';
