+++ src/MOCK_DB_DATA/library-items/table/table-selectable-row.ts
@@ -1,4 +1,4 @@
-import { LibraryItemIds } from 'src/app/views/shared/library-item-ids';
+import { LibraryItemIds } from 'src/app/library-item-ids';
 import { LibraryItem, LibraryItemType } from 'src/app/views/shared/models/library-item.model';
 
 export const TableSelectableRow: LibraryItem[] = [{
