+++ src/app/views/shared/dynamic-library-examples/dynamic-library-example.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { LibraryItemIds } from '../library-item-ids';
+import { LibraryItemIds } from '../../../library-item-ids';
 
 @Component({
   selector: 'app-dynamic-library-example',
