+++ src/app/views/shared/dynamic-library-examples/ng-bootstrap-examples/accordion-example/accordion-example.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { LibraryItemIds } from '../../../library-item-ids';
+import { LibraryItemIds } from '../../../../../library-item-ids';
 
 @Component({
   selector: 'app-accordion-example',
